Add playGame() playing a game with rule of Bof5
The playGame() function consists of 5 rounds and also calculates the
overall winner of the party. The results can be a win, a lose or a tie
for the player.
The playRound() function had to be modified, instead of returning a
string containing a message, it now returns the winner itself as a
string, except if the game result in a tie, then it will still return a
message.